refactoring initiate session callbacks
