+++ handlers/InitiateSalesforceSessionHandler.ts
@@ -1,10 +1,9 @@
-import { IHttp, IHttpRequest, IModify, IPersistence, IRead } from '@rocket.chat/apps-engine/definition/accessors';
-import { IDepartment, ILivechatMessage, ILivechatRoom, ILivechatTransferData, IVisitor } from '@rocket.chat/apps-engine/definition/livechat';
+import { IHttp, IModify, IPersistence, IRead } from '@rocket.chat/apps-engine/definition/accessors';
+import { ILivechatMessage, ILivechatRoom, IVisitor } from '@rocket.chat/apps-engine/definition/livechat';
 import { IMessage } from '@rocket.chat/apps-engine/definition/messages';
-import { RocketChatAssociationModel, RocketChatAssociationRecord } from '@rocket.chat/apps-engine/definition/metadata';
 import { IUser } from '@rocket.chat/apps-engine/definition/users';
 import { getServerSettingValue, sendDebugLCMessage, sendLCMessage } from '../helperFunctions/GeneralHelpers';
-import { getAuthTokens, setBotStatus } from '../helperFunctions/RocketChatAPIHelpers';
+import { checkCurrentChatStatus } from '../helperFunctions/InitiateSalesforceSessionHelpers/CheckCurrentStatusHelper';
 import { getSessionTokens, pullMessages, sendChatRequest } from '../helperFunctions/SalesforceAPIHelpers';
 import { checkForErrorEvents, checkForEvent } from '../helperFunctions/SalesforceMessageHelpers';
 
@@ -13,7 +12,6 @@ export class InitiateSalesforceSession {
 
 	public async exec() {
 		const salesforceBotUsername: string = (await this.read.getEnvironmentReader().getSettings().getById('salesforce_bot_username')).value;
-
 		const lmessage: ILivechatMessage = this.message;
 		const lroom: ILivechatRoom = lmessage.room as ILivechatRoom;
 		const LcAgent: IUser = lroom.servedBy ? lroom.servedBy : this.message.sender;
@@ -56,7 +54,6 @@ export class InitiateSalesforceSession {
 		const LcVisitor: IVisitor = lroom.visitor;
 		const LcVisitorName = LcVisitor.name;
 		const LcVisitorEmailsArr = LcVisitor.visitorEmails;
-
 		let LcVisitorEmail: string = 'No email provided';
 		if (LcVisitorEmailsArr) {
 			const t = LcVisitorEmailsArr[0].address;
@@ -64,12 +61,10 @@ export class InitiateSalesforceSession {
 		}
 
 		await sendDebugLCMessage(this.read, this.modify, this.message.room, 'Initiating session with Salesforce', LcAgent);
-
 		await getSessionTokens(this.http, salesforceChatApiEndpoint)
 			.then(async (res) => {
 				console.log('Generating session id, Response:', res);
 				await sendDebugLCMessage(this.read, this.modify, this.message.room, `Session initiated with Saleforce:: ${JSON.stringify(res)}`, LcAgent);
-
 				const { id, affinityToken, key } = res;
 				await sendChatRequest(
 					this.http,
@@ -92,17 +87,13 @@ export class InitiateSalesforceSession {
 							`Chat request to Salesforce: ${JSON.stringify(sendChatRequestres)}`,
 							LcAgent,
 						);
-
 						await pullMessages(this.http, salesforceChatApiEndpoint, affinityToken, key)
 							.then(async (pullMessagesres) => {
 								console.log('Chat request sent, checking for response , Response:', pullMessagesres);
-
 								const pullMessagesContent = pullMessagesres.content;
 								const pullMessagesContentParsed = JSON.parse(pullMessagesContent || '{}');
 								const pullMessagesMessageArray = pullMessagesContentParsed.messages;
-
 								const isChatRequestSuccess = checkForEvent(pullMessagesMessageArray, 'ChatRequestSuccess');
-
 								if (isChatRequestSuccess === true) {
 									const chatSuccessMessageArray = pullMessagesMessageArray[0].message;
 									const { queuePosition } = chatSuccessMessageArray;
@@ -115,7 +106,6 @@ export class InitiateSalesforceSession {
 										await sendLCMessage(this.modify, this.message.room, queuePosMessage, LcAgent);
 									}
 								}
-
 								await sendDebugLCMessage(
 									this.read,
 									this.modify,
@@ -123,125 +113,6 @@ export class InitiateSalesforceSession {
 									`Current request status: ${pullMessagesContentParsed.messages[0].type}`,
 									LcAgent,
 								);
-
-								const checkAgentStatusCallback = async (data?, error?) => {
-									if (error) {
-										console.log('Check whether agent accepted request, Callback error:', error);
-										sendLCMessage(this.modify, this.message.room, error, LcAgent);
-										return;
-									}
-
-									const contentData = data.content;
-									const contentParsed = JSON.parse(contentData || '{}');
-									console.log('Check whether agent accepted request, Callback Response:', contentParsed);
-
-									await getAuthTokens(http, rocketChatServerUrl, salesforceBotUsername, salesforceBotPassword)
-										.then(async (loginRes) => {
-											const { authToken, userId } = loginRes;
-											await setBotStatus(http, rocketChatServerUrl, authToken, userId)
-												.then(async () => {
-													const assoc = new RocketChatAssociationRecord(RocketChatAssociationModel.ROOM, this.message.room.id);
-													const sessionTokens = { id, affinityToken, key };
-													await this.persistence.createWithAssociation(sessionTokens, assoc);
-
-													const roomId = this.message.room.id;
-													const room: ILivechatRoom = (await this.read.getRoomReader().getById(roomId)) as ILivechatRoom;
-													const targetDepartment: IDepartment = (await this.read
-														.getLivechatReader()
-														.getLivechatDepartmentByIdOrName(targetDeptName)) as IDepartment;
-													const transferData: ILivechatTransferData = {
-														currentRoom: room,
-														targetDepartment: targetDepartment.id,
-													};
-													await this.modify.getUpdater().getLivechatUpdater().transferVisitor(LcVisitor, transferData);
-												})
-												.catch(async (statusErr) => {
-													console.log('Setting Salesforce bot status, Error:', statusErr);
-													await sendLCMessage(this.modify, this.message.room, technicalDifficultyMessage, LcAgent);
-													await sendDebugLCMessage(
-														this.read,
-														this.modify,
-														this.message.room,
-														`Error in setting SF bot stauts: ${statusErr}`,
-														LcAgent,
-													);
-												});
-										})
-										.catch(async (loginErr) => {
-											console.log('Performing Salesforce bot login, Error:', loginErr);
-											await sendLCMessage(this.modify, this.message.room, technicalDifficultyMessage, LcAgent);
-											await sendDebugLCMessage(
-												this.read,
-												this.modify,
-												this.message.room,
-												`Error in performing SF bot login: ${loginErr}`,
-												LcAgent,
-											);
-										});
-								};
-
-								const { http, modify, message } = this;
-								async function checkCurrentChatStatus(callback) {
-									pullMessages(http, salesforceChatApiEndpoint, affinityToken, key)
-										.then(async (response) => {
-											if (response.statusCode === 403) {
-												console.log('Check whether agent accepted request, Session Expired. ', response);
-												callback([], 'Chat session expired.');
-												return;
-											} else if (response.statusCode === 204 || response.statusCode === 409) {
-												console.log('Check whether agent accepted request, Empty Response.', response);
-												await checkCurrentChatStatus(callback);
-											} else {
-												console.log('Check whether agent accepted request, response here:', response);
-
-												const { content } = response;
-												const contentParsed = JSON.parse(content || '{}');
-												const messageArray = contentParsed.messages;
-
-												const isQueueUpdate = checkForEvent(messageArray, 'QueueUpdate');
-												if (isQueueUpdate === true) {
-													console.log('isQueueUpdate: ', isQueueUpdate);
-													const queueUpdateMessages = messageArray[0].message;
-													const queueUpdatePosition = queueUpdateMessages.position;
-
-													if (queueUpdatePosition === 1) {
-														const queueEmptyMessage = LAQueueEmptyMessage.replace(/%s/g, queueUpdatePosition);
-														await sendLCMessage(modify, message.room, queueEmptyMessage, LcAgent);
-													} else if (queueUpdatePosition > 1) {
-														const queuePosMessage = LAQueuePositionMessage.replace(/%s/g, queueUpdatePosition);
-														await sendLCMessage(modify, message.room, queuePosMessage, LcAgent);
-													}
-												}
-
-												const isChatAccepted = checkForEvent(messageArray, 'ChatEstablished');
-												if (isChatAccepted === true) {
-													console.log('Chat accepted by agent: ', isChatAccepted);
-													console.log('Check whether agent accepted request, Chat ended by Live Agent.');
-													callback(response);
-													return;
-												} else if (isChatAccepted === false) {
-													console.log('Chat accepted by agent: ', isChatAccepted);
-
-													const isChatRequestFail = checkForEvent(messageArray, 'ChatRequestFail');
-													if (isChatRequestFail === true) {
-														console.log('Chat request fail: ', isChatRequestFail);
-														callback([], technicalDifficultyMessage);
-														return;
-													} else {
-														await checkCurrentChatStatus(callback);
-													}
-												} else {
-													console.log('Check whether agent accepted request, Unresolved Response:', response);
-													await checkCurrentChatStatus(callback);
-												}
-											}
-										})
-										.catch(async (error) => {
-											console.log('Check whether agent accepted request, Error: ', error);
-											await checkCurrentChatStatus(callback);
-										});
-								}
-
 								if (pullMessagesContentParsed.messages[0].type === 'ChatRequestFail') {
 									await checkForErrorEvents(
 										this.read,
@@ -253,7 +124,26 @@ export class InitiateSalesforceSession {
 									);
 								} else {
 									console.log('Chat request sent, checking for response, Executing Function:');
-									checkCurrentChatStatus(checkAgentStatusCallback);
+									await checkCurrentChatStatus(
+										this.http,
+										this.modify,
+										this.persistence,
+										this.message,
+										this.read,
+										salesforceChatApiEndpoint,
+										rocketChatServerUrl,
+										salesforceBotUsername,
+										salesforceBotPassword,
+										id,
+										affinityToken,
+										key,
+										targetDeptName,
+										LcAgent,
+										LAQueueEmptyMessage,
+										LAQueuePositionMessage,
+										technicalDifficultyMessage,
+										LcVisitor,
+									);
 								}
 							})
 							.catch(async (error) => {
